Clean up hasty commit to ocrla-2018p.ipynb
Earlier, work was in progress to display the data live in the notebook,
in HTML rather than running analyze_round.py external to the notebook
and displaying the console output, But it was committed too early.

There's more to do to properly display it, but this gives a hint
while still showing it the original way.